+++ crates/store/src/lib.rs
@@ -1,17 +1,19 @@
 use std::time::Duration;
 
 mod blocks;
-pub mod db;
-pub mod errors;
-pub mod genesis;
+mod db;
+mod errors;
+mod genesis;
 mod nullifier_tree;
-pub mod server;
-pub mod state;
+mod server;
+mod state;
+
+pub use genesis::GenesisState;
+pub use server::{DataDirectory, Store};
 
 // CONSTANTS
 // =================================================================================================
-pub const COMPONENT: &str = "miden-store";
-pub const GENESIS_STATE_FILENAME: &str = "genesis.dat";
+const COMPONENT: &str = "miden-store";
 
 /// Number of sql statements that each connection will cache.
 const SQL_STATEMENT_CACHE_CAPACITY: usize = 32;